# docs/swagger.yaml
basePath: /api
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        description: machine-friendly error code
        example: FORBIDDEN
        type: string
      error:
        description: always true for error cases
        example: true
        type: boolean
      message:
        description: human-readable error message
        example: Forbidden
        type: string
    type: object
...
  models.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        description: field name -> list of error messages
        type: object
      message:
        example: Validation failed
        type: string
    type: object
...
  quotes.MyQuoteItem:
    properties:
      amount_cents:
        type: integer
      case_category:
        description: optional
        type: string
      case_id:
        type: string
      case_status:
        description: optional
        type: string
      case_title:
        description: from cases.title
        type: string
      created_at:
        type: string
      days:
        type: integer
      id:
        type: string
      note:
        type: string
      status:
        type: string
    type: object
  quotes.PageMyQuotes:
    properties:
      items:
        items:
          $ref: '#/definitions/quotes.MyQuoteItem'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  quotes.UpsertQuoteRequest:
    properties:
      amount_cents:
        description: min S$10, max S$1,000,000
        maximum: 100000000
        minimum: 1
        type: integer
      case_id:
        type: string
      days:
        maximum: 365
        minimum: 1
        type: integer
      note:
        maxLength: 500
        type: string
    required:
    - amount_cents
    - case_id
    - days
    type: object
info:
  contact:
    email: aldoetobex@gmail.com
    name: Aldo Rifki Putra
  description: 'API for a mini legal marketplace: clients post cases, lawyers submit
    quotes, clients accept & pay, and lawyers access files via signed URLs.'
  title: Mini Legal Marketplace API
  version: "1.0"
paths:
...
  /cases/{id}/quotes:
    get:
      description: Client owner sees all quotes for their case (paginated)
      parameters:
      - description: case id (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quotes.PageMyQuotes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Quotes by case (owner)
      tags:
      - quotes
...
  /quotes:
    post:
      consumes:
      - application/json
      description: Lawyer creates or updates a quote while the case is still OPEN
      parameters:
      - description: Quote upsert payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/quotes.UpsertQuoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: id, status, amount_cents, days, note
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: immutable or case not open
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Submit or update a quote (1 active per case per lawyer)
      tags:
      - quotes
  /quotes/mine:
    get:
      description: Lawyer lists their quotes (filter by status, with pagination).
        Includes case title.
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        type: integer
      - description: proposed|accepted|rejected
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/quotes.PageMyQuotes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my quotes
      tags:
      - quotes
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: 'Format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"